debian/*: adding a starter wrapper script
On systems like Wheezy we have also the openjdk-6-jre package that can
be set up as the system default. The gridvis-desktop package needs the
openjdk-7-jre. So we can not trust that /usr/lib/jvm/default-java points
to the version 7 of the openjdk package, we have to call the
gridvis-desktop starter with a correct $jdkhome value to ensure the
gridvis-desktop can work correct. To ensure this there is now a starting
wrapper script within the $PATH so no explicit knowledge about the
starting script for gridvis-desktop is needed on the user side.