gridvis-desktop.preinst: adding installer helper
The gridvis-desktop is creating some caching informations while a first
startup. In case of a package update the new gridvis-desktop package
would use these old cached informations, this is a unwanted behaviour.
In case of a package update we have to move out this cached informations
of the way. Doing this by using preinst helper mechanism.